Provide dependencies using koin & create generic viewModel observable. 🥳
