+++ ios/tv-maniac.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -5,9 +5,9 @@
         "package": "TvManiac",
         "repositoryURL": "https://github.com/c0de-wizard/tvmaniac-swift-packages",
         "state": {
-          "branch": null,
-          "revision": "bffe3f077c43761beddf58bab8e8c48a670b72b0",
-          "version": "0.2.0"
+          "branch": "main",
+          "revision": "84066f1e7c25d39a0b20d3836affa81f958beb7a",
+          "version": null
         }
       }
     ]
